<@rfelipe,@brendamoniqueb>arruma dados editaveis

+++ app/src/main/java/com/thoughtworks/aceleradora/solicitacao/controller/SolicitacaoController.java
@@ -96,7 +96,7 @@ public class SolicitacaoController {
             solicitacaoAtu.setEndereco(solicitacao.getEndereco());
             solicitacaoAtu.setSituacao(solicitacao.getSituacao());
             solicitacaoAtu.setGenero(solicitacao.getGenero());
-            solicitacaoAtu.setPeso(solicitacao.getPeso());
+            solicitacaoAtu.setPeso((Float)solicitacao.getPeso());
             solicitacaoAtu.setDataNascimento(solicitacao.getDataNascimento());
             solicitacaoAtu.setDataTransplante(solicitacao.getDataTransplante());
